sales invoice creation - timesheets
 - make sure timesheets get summarized properly

Reviewed-on: http://gerrit.tine20.com/customers/630
Tested-by: Jenkins CI (http://ci.tine20.com/)

+++ tine20/Sales/Model/InvoicePosition.php
@@ -79,7 +79,7 @@ class Sales_Model_InvoicePosition extends Tinebase_Record_Abstract
             ),
             'quantity' => array(
                 'label' => 'Quantity', //_('Quantity')
-                'type'  => 'integer',
+                'type'  => 'float',
                 'summaryType' => 'sum',
             ),
         )
